% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_presentation.R
\name{create_presentation}
\alias{create_presentation}
\title{Title}
\usage{
create_presentation(package, file = "")
}
\arguments{
\item{package}{A file path to the root directory of an R package source
folder.}

\item{file}{The file name for your .qmd file. This can be a path so long as
it ends with a `file_name.qmd`}
}
\value{
This function creates and renders a .qmd presentation but does not
  return an R object.
}
\description{
Title
}
