% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{create_yaml}
\alias{create_yaml}
\title{Create `_pkgslides.yml`}
\usage{
create_yaml(package)
}
\arguments{
\item{package}{A file path to the root directory of an R package source
folder.}
}
\value{
This will not fill the file, simply create it.
}
\description{
Create `_pkgslides.yml`
}
