% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{.check_yaml}
\alias{.check_yaml}
\title{Fill an Incomplete `_pkgslides.yml`}
\usage{
.check_yaml(yaml)
}
\arguments{
\item{yaml}{}
}
\value{
A list representing a yaml file
}
\description{
Fill an Incomplete `_pkgslides.yml`
}
\keyword{internal}
