% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{.choose_datasets}
\alias{.choose_datasets}
\title{Title}
\usage{
.choose_datasets(p, datasets)
}
\arguments{
\item{p}{A list of roxygen2 blocks}

\item{datasets}{A list from `_pkgslides.yml`}
}
\value{
A list of r
}
\description{
Title
}
\keyword{internal}
