% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{.choose_slides}
\alias{.choose_slides}
\title{Title}
\usage{
.choose_slides(p, slides)
}
\arguments{
\item{p}{A list of roxygen2 blocks}

\item{slides}{A list from `_pkgslides.yml`}
}
\value{
A list of roxygen2 blocks
}
\description{
Title
}
\keyword{internal}
