% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_datasets.R
\name{.collate_datasets}
\alias{.collate_datasets}
\title{Title}
\usage{
.collate_datasets(dataset_details)
}
\arguments{
\item{dataset_details}{A list of function details from `.get_dataset_tags`}
}
\value{
A character vector of properties formatted for writing to a file
}
\description{
Title
}
\keyword{internal}
