% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_description.R
\name{.collate_description}
\alias{.collate_description}
\title{Collate Description}
\usage{
.collate_description(package_details, chunk_opt)
}
\arguments{
\item{package_details}{A list of function details from `.get_description`}
}
\value{
A character vector of properties formatted for writing to a file
}
\description{
Collate Description
}
\keyword{internal}
