% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.collate_functions}
\alias{.collate_functions}
\title{Title}
\usage{
.collate_functions(function_details)
}
\arguments{
\item{function_details}{A list of function details from `.get_function_tags`}
}
\value{
A character vector of properties formatted for writing to a file
}
\description{
Title
}
\keyword{internal}
