% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_title.R
\name{.collate_title}
\alias{.collate_title}
\title{Title}
\usage{
.collate_title(title_details)
}
\arguments{
\item{package_details}{A list of function details from `.get_title`}
}
\value{
A character vector of properties formatted for writing to a file
}
\description{
Title
}
\keyword{internal}
