% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.construct_verticals}
\alias{.construct_verticals}
\title{Title}
\usage{
.construct_verticals(r, r_files, function_contents, chunk_opt)
}
\arguments{
\item{r}{The R source file for which a vertical is created}

\item{r_files}{A vector of unique R source file names}

\item{function_contents}{A list of roxygen2 blocks for the `r` file}

\item{chunk_opt}{"eval" or "echo" for a chunk option. This is
determined by the install status of the chosen package.}
}
\value{
A character vector of properties formatted for writing to a file
}
\description{
Title
}
\keyword{internal}
