% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.drop_function_options}
\alias{.drop_function_options}
\title{Title}
\usage{
.drop_function_options(item, item_name, yaml)
}
\arguments{
\item{item}{A list item from `.get_tag_list`}

\item{item_name}{The name of a list item from `.get_tag_list`}

\item{yaml}{A list of properties from `.parse_yaml()`}
}
\value{
Either `item` or `NULL`
}
\description{
Title
}
\keyword{internal}
