% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.drop_null_from_list}
\alias{.drop_null_from_list}
\title{Title}
\usage{
.drop_null_from_list(l)
}
\arguments{
\item{l}{A list}
}
\value{
A list with null values removed
}
\description{
Title
}
\keyword{internal}
