% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.drop_options}
\alias{.drop_options}
\title{Drop Options}
\usage{
.drop_options(item, item_name, yaml, type)
}
\arguments{
\item{item}{A list item from `.get_*_tags`}

\item{item_name}{The name of a list item from `.get_*_tags`}

\item{yaml}{A list of properties from `.parse_yaml()`}
}
\value{
Either `item` or `NULL`
}
\description{
Drop Options
}
\keyword{internal}
