% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{.filter_roxygen_blocks}
\alias{.filter_roxygen_blocks}
\title{Title}
\usage{
.filter_roxygen_blocks(p, y_names)
}
\arguments{
\item{p}{A list of roxygen2 blocks}

\item{y_names}{A list of yaml slide option name}
}
\value{
A list of roxygen2 blocks
}
\description{
Title
}
\keyword{internal}
