% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{.filter_roxygen_exported}
\alias{.filter_roxygen_exported}
\title{Title}
\usage{
.filter_roxygen_exported(p, exported = TRUE)
}
\arguments{
\item{p}{A list of roxygen2 blocks}

\item{exported}{A boolean if you want exported or non-exported functions
returned}
}
\value{
A list of roxygen2 blocks
}
\description{
Title
}
\keyword{internal}
