% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{.filter_roxygen_functions}
\alias{.filter_roxygen_functions}
\title{Title}
\usage{
.filter_roxygen_functions(y, y_names, p)
}
\arguments{
\item{y}{A list of yaml slide options}

\item{y_names}{A list of yaml slide option name}

\item{p}{A list of roxygen2 blocks}
}
\value{
A list of roxygen2 blocks
}
\description{
Title
}
\keyword{internal}
