% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_presentation.R
\name{.find_file}
\alias{.find_file}
\title{Title}
\usage{
.find_file(package, file)
}
\arguments{
\item{package}{A path to a package source}

\item{file}{A file path}
}
\value{
A file path to write the .qmd to
}
\description{
Title
}
\keyword{internal}
