% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.fit_content}
\alias{.fit_content}
\title{Puts Content in an `r-fit-content` div}
\usage{
.fit_content(content)
}
\arguments{
\item{content}{A length-one character vector}
}
\value{
A length-one character vector that renders a Quarto `r-fit-content`
  div
}
\description{
Puts Content in an `r-fit-content` div
}
\keyword{internal}
