% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.get_credits}
\alias{.get_credits}
\title{Get Credits}
\usage{
.get_credits(package)
}
\arguments{
\item{package}{A file path to the root directory of an R package source
folder, the name of an R package from CRAN, or an R package from
GitHub in the username/repository format.}
}
\value{
A character vector of author and maintainer details
}
\description{
Get Credits
}
\keyword{internal}
