% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_datasets.R
\name{.get_dataset_tags}
\alias{.get_dataset_tags}
\title{Get Dataset Tags}
\usage{
.get_dataset_tags(block, yaml)
}
\arguments{
\item{block}{A roxygen2 block}

\item{yaml}{A list of properties from `.parse_yaml()`}
}
\value{
A list of roxygen2 tags for a specific source file
}
\description{
Get Dataset Tags
}
\keyword{internal}
