% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_datasets.R
\name{.get_datasets}
\alias{.get_datasets}
\title{Title}
\usage{
.get_datasets(datasets, yaml)
}
\arguments{
\item{yaml}{A list of properties from `.parse_yaml()`}

\item{file}{A file path}
}
\value{
A list of details ready to be collated into slides
}
\description{
Title
}
\keyword{internal}
