% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.get_desc}
\alias{.get_desc}
\title{Get a DESCRIPTION Property}
\usage{
.get_desc(package, tag)
}
\arguments{
\item{package}{A file path to the root directory of an R package source
folder.}

\item{tag}{The name of the tag whose contents are desired}
}
\value{
A length-one character vector with the desired text from a package's
  DESCRIPTION file
}
\description{
Get a DESCRIPTION Property
}
\keyword{internal}
