% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.get_function_tags}
\alias{.get_function_tags}
\title{Gets a List of Roxygen2 Tags}
\usage{
.get_function_tags(block, yaml)
}
\arguments{
\item{block}{A roxygen2 block}

\item{yaml}{A list of properties from `.parse_yaml()`}
}
\value{
A list of roxygen2 tags for a specific source file
}
\description{
Gets a List of Roxygen2 Tags
}
\keyword{internal}
