% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{.get_roxygen}
\alias{.get_roxygen}
\title{Title}
\usage{
.get_roxygen(package, yaml)
}
\arguments{
\item{package}{A file path to the root directory of an R package source
folder.}

\item{yaml}{A list of properties from `.parse_yaml()`}
}
\value{
A list of roxygen2 blocks
}
\description{
Title
}
\keyword{internal}
