% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.get_tag}
\alias{.get_tag}
\title{Title}
\usage{
.get_tag(block, tag)
}
\arguments{
\item{block}{A roxygen2 block}

\item{tag}{A character vector of roxygen2 block tag names}
}
\value{
Returns the raw text of the desired tag
}
\description{
Title
}
\keyword{internal}
