% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.get_tag_list}
\alias{.get_tag_list}
\title{Gets a List of Roxygen2 Tags}
\usage{
.get_tag_list(block)
}
\arguments{
\item{block}{A roxygen2 block}
}
\value{
A list of roxygen2 tags for a specific source file
}
\description{
Gets a List of Roxygen2 Tags
}
\keyword{internal}
