% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.get_tags}
\alias{.get_tags}
\title{Title}
\usage{
.get_tags(block)
}
\arguments{
\item{block}{A roxygen2 block}
}
\value{
A list of roxygen2 properties for a specific source file
}
\description{
Title
}
\keyword{internal}
