% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{.organize_datasets}
\alias{.organize_datasets}
\title{Title}
\usage{
.organize_datasets(p)
}
\arguments{
\item{p}{A list of roxygen2 blocks}
}
\value{
A list of roxygen2 blocks
}
\description{
Title
}
\keyword{internal}
