% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{.organize_functions}
\alias{.organize_functions}
\title{Title}
\usage{
.organize_functions(p)
}
\arguments{
\item{p}{A list of roxygen2 blocks}
}
\value{
The list p in the same file order, but functions are organized by
  exported first
}
\description{
Title
}
\keyword{internal}
