% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{.parse_yaml}
\alias{.parse_yaml}
\title{Parse `_pkgslides.yml`}
\usage{
.parse_yaml(package)
}
\arguments{
\item{package}{A file path to the root directory of an R package source
folder.}
}
\value{
A list representing a yaml file
}
\description{
Parse `_pkgslides.yml`
}
\keyword{internal}
