% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{.process_choose_functions}
\alias{.process_choose_functions}
\title{Title}
\usage{
.process_choose_functions(yaml, choose)
}
\arguments{
\item{yaml}{A list representing a yaml file}

\item{choose}{The list passed to `choose_functions` in the `create_yaml`
function}
}
\value{
A list representing a yaml file
}
\description{
Title
}
\keyword{internal}
