% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.process_params}
\alias{.process_params}
\title{Title}
\usage{
.process_params(param)
}
\arguments{
\item{param}{A function parameter from an roxygen2 tag}
}
\value{
A length-one character vector of parameters formatted as a bulleted
  list for writing to a file
}
\description{
Title
}
\keyword{internal}
