% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{.set_as_true}
\alias{.set_as_true}
\title{Set Empty Values as True}
\usage{
.set_as_true(yaml, type, options)
}
\arguments{
\item{yaml}{A yaml file as a list}

\item{type}{"functions" or "datasets"}

\item{options}{A vector of the options exposed to users in the yaml format}
}
\value{
A yaml where non-included options are set to TRUE
}
\description{
Set Empty Values as True
}
\keyword{internal}
