% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_presentation.R
\name{build_presentation}
\alias{build_presentation}
\title{Create a Revealjs Presentation with Quarto}
\usage{
build_presentation(package = getwd(), file = NULL, yaml = create_yaml())
}
\arguments{
\item{package}{A file path to the root directory of an R package source
folder, the name of an R package from CRAN, or an R package from
GitHub in the username/repository format.}

\item{file}{The file name for your .qmd file. This can be a path so long as
it ends with a `file_name.qmd`}

\item{yaml}{A `_pkgslides.yml` file or function call to `create_yaml()`}
}
\value{
This function creates and renders a .qmd presentation but does not
  return an R object.
}
\description{
When pointed at the root directory of an R Package Project, this
  function will automatically generate and render a Quarto/Revealjs
  presentation for that package. It will have a title slide and description
  slide generated from the package's DESCRIPTION file and each exported
  function will have slides with its description and returns, parameters,
  examples, and source code.
}
