% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{create_yaml}
\alias{create_yaml}
\title{Create `_pkgslides.yml`}
\usage{
create_yaml(
  path = getwd(),
  format_theme = c(),
  format_functions = list(),
  format_datasets = list(),
  choose_functions = list(),
  choose_datasets = c()
)
}
\arguments{
\item{path}{A file path to where you want the yaml file written. This should
not end in a slash of any kind.}

\item{format_theme}{A length-one character vector of theme details}

\item{format_functions}{A named list of format function options}

\item{format_datasets}{A named list of format dataset options}

\item{choose_functions}{A list of file or function names}

\item{choose_datasets}{A vector of dataset names}
}
\value{
This will create the file and write to it, then return the
 file path.
}
\description{
Create `_pkgslides.yml`
}
