% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_roxygen.R
\name{.filter_roxygen_custom}
\alias{.filter_roxygen_custom}
\title{Title}
\usage{
.filter_roxygen_custom(p, slides)
}
\arguments{
\item{p}{A list of roxygen2 blocks}

\item{slides}{}
}
\value{
A list of roxygen2 blocks
}
\description{
Title
}
\keyword{internal}
