% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_presentation.R
\name{.find_package}
\alias{.find_package}
\title{Title}
\usage{
.find_package(package)
}
\arguments{
\item{package}{A package name or file path to a package source}
}
\value{
A path to a package source
}
\description{
Title
}
\keyword{internal}
