% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_description.R
\name{.get_description}
\alias{.get_description}
\title{Title}
\usage{
.get_description(package)
}
\arguments{
\item{package}{A file path to the root directory of an R package source
folder, the name of an R package from CRAN, or an R package from
GitHub in the username/repository format.}
}
\value{
A list of roxygen2 properties for the package's DESCRIPTION file used
  to create the package description slide
}
\description{
Title
}
\keyword{internal}
