% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.get_file_from_path}
\alias{.get_file_from_path}
\title{Get File From Path}
\usage{
.get_file_from_path(path)
}
\arguments{
\item{path}{A file path}
}
\value{
A length-one character vector with the last portion of the supplied
  path
}
\description{
Get File From Path
}
\keyword{internal}
