% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.get_functions}
\alias{.get_functions}
\title{Get Functions}
\usage{
.get_functions(functions, yaml)
}
\arguments{
\item{functions}{A list of roxygen2 blocks for functions}

\item{yaml}{A list of properties from `.parse_yaml()`}
}
\value{
A list of details ready to be collated into slides
}
\description{
Get Functions
}
\keyword{internal}
