% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_functions.R
\name{.get_functions}
\alias{.get_functions}
\title{Title}
\usage{
.get_functions(package, yaml)
}
\arguments{
\item{yaml}{A list of properties from `.parse_yaml()`}

\item{file}{A file path}
}
\value{
A character vector of properties formatted for writing to a file
}
\description{
Title
}
\keyword{internal}
