% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{.parse_yaml}
\alias{.parse_yaml}
\title{Parse `_pkgslides.yml`}
\usage{
.parse_yaml(file)
}
\arguments{
\item{file}{A file path the _pkgslides.yml file}
}
\value{
A list representing a yaml file
}
\description{
Parse `_pkgslides.yml`
}
\keyword{internal}
